.github/build-script/requirements.txt
requests
jinja2
pyyaml